# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API HỆ THỐNG CỬA HÀNG ĐIỆN THOẠI PHÚC MOBILE",
        "description": "Hệ thống API quản lý cửa hàng kinh doanh điện thoại di động và phụ kiện công nghệ PHÚC MOBILE. Hỗ trợ đầy đủ phân quyền người dùng (RBAC) và chuẩn hóa dữ liệu quan hệ 3NF.\n *     \n *     ### Danh sách Vai trò (Roles):\n *     - **superadmin**: Quyền cao nhất, quản lý toàn bộ hệ thống bao gồm cả nhân viên và khách hàng.\n *     - **admin**: Quản lý nghiệp vụ chính (Điện thoại, Hãng sản xuất, Danh mục, Mã giảm giá, Đơn hàng, Đánh giá).\n *     - **staff**: Nhân viên vận hành cửa hàng, xử lý trạng thái đơn hàng và phê duyệt đánh giá sản phẩm.\n *     - **user**: Khách hàng (Xem sản phẩm, thêm giỏ hàng, đặt mua điện thoại và gửi đánh giá).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server (Tự động nhận diện Host)"
        }
    ],
    "paths": {
        "/admin/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Đăng nhập Admin",
                "description": "API cho phép admin đăng nhập vào hệ thống bằng email và mật khẩu. Nếu thông tin hợp lệ, hệ thống sẽ trả về JWT token để sử dụng cho các API yêu cầu xác thực.",
                "operationId": "adminLogin",
                "requestBody": {
                    "description": "Thông tin đăng nhập của admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email đăng nhập của admin",
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "description": "Mật khẩu đăng nhập",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJh..."
                                        },
                                        "admin": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Admin"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@gmail.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-01-01T10:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-01-01T10:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Sai email hoặc mật khẩu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email hoặc mật khẩu admin không đúng"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu gửi lên không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The email field is required."
                                                ],
                                                "password": [
                                                    "The password field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/logout": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Đăng xuất Admin",
                "operationId": "62127654218171e947bfb021ecd6df28",
                "responses": {
                    "200": {
                        "description": "Đăng xuất thành công"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/me": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lấy thông tin Admin đang đăng nhập",
                "operationId": "5894139af62b9ea3a13677699d1c941b",
                "responses": {
                    "200": {
                        "description": "Lấy thông tin thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "admin": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Admin"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@gmail.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "superadmin"
                                                },
                                                "is_active": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/staff": {
            "get": {
                "tags": [
                    "Staff Management"
                ],
                "summary": "Lấy danh sách admin/nhân viên",
                "operationId": "f3488f07388434bbe68ff61fbca1f661",
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "role": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    },
                    "403": {
                        "description": "Không có quyền"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Staff Management"
                ],
                "summary": "Tạo admin/nhân viên mới",
                "operationId": "cb053e32c7468bba13f4a31fc1a6e5ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Staff"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "staff@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "staff"
                                        ],
                                        "example": "staff"
                                    },
                                    "is_active": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tạo tài khoản quản trị thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "role": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập JWT token vào đây để thực hiện các yêu cầu có bảo mật.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}